refactor: fix sign-compare warnings

+++ src/fileio/scan.c
@@ -60,7 +60,7 @@ static ssize_t readline(char** lineptr, size_t* n, FILE* stream) {
     }
     p = bufptr;
     while (c != EOF) {
-        if ((p - bufptr) > (size - 1)) {
+        if ((ssize_t)(p - bufptr) > (ssize_t)(size - 1)) {
             size = size + 128;
             bufptr = realloc(bufptr, size);
             if (bufptr == NULL) {
